[#82] Update tests
